+++ lib/ui/i2i_tab/view_models/i2i_tab_viewmodel.dart
@@ -6,4 +6,5 @@ import 'package:nai_casrand/data/models/payload_config.dart';
 class I2iTabViewmodel extends ChangeNotifier {
   ParamConfig get paramConfig => GetIt.I<PayloadConfig>().paramConfig;
   bool get isV4 => paramConfig.model.contains('-4-');
+  bool get isV5 => paramConfig.model.contains('-5-');
 }